Improve GN value origins for error messages.
Sets the origin for a value when the value is dereferenced. Previously if you did
  foo = 6
  deps = [ foo ]
You would get the message:
  ERROR at //BUILD.gn:15:9: This is not a string.
    foo = 6
          ^
Which in this case is pretty confusing since that's obviously the case, and you really want to know where it was dereferenced.

With this patch the message is:
  ERROR at //BUILD.gn:21:5: This is not a string.
      foo,
      ^--
  Instead I see a integer = 6
which is more actionable.

R=ajwong@chromium.org

Review URL: https://codereview.chromium.org/439513003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea4a89da7b5677dd1ed774b0719745338dbe9613