Reverse the order of GN's cycle printing.
When printing dependency cycles, the old meticulously walked the list backwards and explained that this was because the list was in the reverse order of the dependencies.

This is wrong, the list is in the normal forward direction and doing that prints out the reverse cycle which is very confusing. Reverse the order of printing to get the correct cycle.

Review URL: https://codereview.chromium.org/557103003

Cr-Original-Commit-Position: refs/heads/master@{#294436}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f42d9b7b286f30e793f0d57f814f281db232844