Enhance GN introspection
This adds additional matching and output capabilities to the various GN introspection commands (desc, ls, refs). It adds --as for controlling output, and --testonly and --type for controlling which ones to show.

It adds support for matching multiple input arguments for the refs, ls, and check commands.

It adds support for finding which targets reference a config in the refs command.

Review URL: https://codereview.chromium.org/937003002

Cr-Original-Commit-Position: refs/heads/master@{#317159}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e604e508f8b3b08d89b5e98ef1d376c74d471a38