Count data deps even if it was already a regular dep
Previously GN's runtime deps computation would only count a target once, but
this ignores the fact that whether a target is a data dependency or a regular
one affects how it's added.

This patch revisits a target if it's depended on in a data dependency
even if it's already been seen as a regular dependency.

Enhance the runtime deps documentation.

Review URL: https://codereview.chromium.org/1250883002

Cr-Original-Commit-Position: refs/heads/master@{#339778}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 295e07207149d215f749dffb68c9c56e1188bc4d