Restore clang-format formatting.
As discussed in https://codereview.chromium.org/2556743002/,
formatting is strange "because clang-format tries to align
columns, even if there's only one row with aligned columns.
I agree that this looks weird, and it sounds like it will be
fixed."

Removing the "// clang-format off" section as it has a cost
since it requires developer to manually code format the region.

BUG=None

Review-Url: https://codereview.chromium.org/2563593004
Cr-Original-Commit-Position: refs/heads/master@{#437521}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3a5f1623bc298bb5285819d285b37cc5da06e598