Tweak flags to improve size/speed for posix
I noticed the GN cipd binary was ~7M instead of the ~1-2M it ought to
be, so update to strip the release binary. Additionally, turn on -flto
since we build with clang anyway, which makes generation of the Chrome
ninja files about 5% faster on my Linux box.

Change-Id: Ib389b20e6729a74b922fc7a3ce5b4412a3972de9
Reviewed-on: https://gn-review.googlesource.com/1820
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Scott Graham <scottmg@chromium.org>