Fix gn_unittests
Scope::KeyValueMap used to be a std::unordered_map, which was confusing the
newly added output conversion unittests that uses string comparison to check if
two values are the same.  The fact that the outputs were unordered means that,
depending on the version of the C++ library used, a value might convert to
"a=0\nb=1" or "b=1\na=0".

This CL makes Scope::KeyValueMap an ordered map instead.

R=brettw@google.com
CC=juliehockett@google.com

Change-Id: Ia3dd1b6a01759f94d1edc96c0bd2a115f743c8c7
Reviewed-on: https://gn-review.googlesource.com/2663
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>