Minor fixes to tracing
The chrome://tracing expects the timestamp to be in microseconds, but
it had been stored in the TraceItem in nanoseconds, so this fixes it.

Also adds quotes around the thread_id, since on Mac it is written in hex
format and so the JSON parser doesn't recognize it.

Finally, adds toolchain information to the import traces.

Change-Id: If5960074d4fc8301cbcc4a3524cb17ab6a95010a
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/5140
Commit-Queue: Julie Hockett <juliehockett@google.com>
Reviewed-by: Scott Graham <scottmg@chromium.org>