Define string16 in terms of std::u16string.
This can be done now that we have C++17 support. The typedef is left for
simplicity. It will be replaced in a separate pass.

Remove wstring completely. On Windows we now use explicit 16-bit strings
and cast to make syscalls work. A new helper is added to support this.

Remove some unnecessary conversion functions and NACL defines.

Change-Id: I30c5b35c31e59510474fff71dc7256fb56cda641
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/6020
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Scott Graham <scottmg@chromium.org>