Handle metadata walk_keys relative to the target directory
Don't process relative labels relative to root, process them relative
to the target directory.

To illustrate this, take the following example:

//x:x
//x:meta
//x/z:z

Where x is a target that depends on z, and meta is a generated_file
target that sets walk_keys to just "z", the collection would fail
because generated_file would try to resolve "z" (which is a relative
path) against the root (i.e. as //z:z). This patches changes the logic
to always resolve the walk_keys against the target current directory,
i.e. in case of //x:meta it'll be resolved against //x which is correct
since "z" is relative to that.

Bug: 117
Change-Id: I4da2a3d5e279a970794cdf163f596d8c9a107f4b
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/6380
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>