Improve support for frameworks in gn
Add "frameworks" and "framework_dirs" variables to targets and configs
that are propagated using the following rules:

-   "frameworks" is propagated to all linkable targets (executables,
    shared libraries and complete static libraries)

-   "framework_dirs" is propagated to direct compile dependencies as
    "include_dirs" and also to all linkable targets (executables,
    shared libraries and complete static libraries)

The non-conventional propagation rules for "framework_dirs" is there
to support how framework are used on Apple's platforms (macOS & iOS).
A framework bundle is a way to ship a shared library with its public
headers, so the path to the framework needs to be propagated to the
compiler (to resolve import statement for the framework's header) and
to the linker (to find the path to the shared library contained in the
framework).

Add target of type CREATE_BUNDLE that generate a framework (identified
by the product type "com.apple.product-type.framework") as inherited
libraries since they contains shared library and add them to the list
of implicit dependencies to force the regeneration of linkable targets
if any of the framework API changes (this is bit pessimistic since it
can force an unnecessary rebuild if the linkable target does not link
with the framework, but this should be rare).

Bug: crbug.com/gn/119, crbug.com/1037607
Change-Id: Icef50c38ac02afab71f2348fe3f73b6124a29112
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/6660
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>