Handle an empty interpreter_path in exec_script
If the interpreter_path is the empty string, exec_script() attempts to
construct a command line with the first argument being the empty string
and the second being the name of the program. This fails, where the
user likely expected their script to execute directly.

Change-Id: I70a9f28a3e256753666016c951daafa00882142b
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7460
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>