Fix crash in "gn gen --ide=xcode" when project is malformed
As found while investigating https://crbug.com/1056925, if the
project was malformed, "gn gen --ide=xcode" could crash due to
some undefined behaviour (likely dereferencing an uninitialised
pointer).

This was due to FindApplicationTargetByName() using DCHECK to
enforce a property of the project when said property is not
under the control of the programmer but under the control of
the user writing the project's BUILD.gn files.

Instead change the function implementation to return an error
if the expected property is not verified. That error is then
propagated back to the "gn gen" implementation and displayed
to the user.

All other uses of DCHECK/NOTREACHED in xcode_writer.cc have
been inspected to verify they correspond to assertion about
the state of the code and do not depend on the properties of
the project's BUILD.gn files.

Bug: 150
Change-Id: Ibe711f2d380ef0ed5dca637799563223c248ec7d
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7560
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>