[Xcode] Implement custom sort of the elements in Xcode project
Add way to sort the elements when adding them to PBXGroup and
implement a custom sorting rule that keep folder before files.

Also add helper method to create a child, add it to PBXGroup
and return the value as a pointer to the correct type.

Remove always empty group "Build" from the project.

Bug: none
Change-Id: I7bc1ead9e724879603beb47da9970662ffbc8171
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7641
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>