Add an "outputs" command.
"gn outputs ..." is a new command-line tool that will list the outputs
for a target or source file. It is mostly intended to help users find
how to express a compile command that just compiles what they're working
on (often the output file to write is difficult to figure out).

Do some refactoring work to share this computation with
get_target_outputs and the "gn desc" code (both of which duplicated it
before).

Change-Id: I5b4a0ef30b106598900a96d9eee1ba516f7ac1f7
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7740
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Scott Graham <scottmg@chromium.org>