format: Fix formatting not taking into account suffixes
In subprinters (evaluated for penalty values), suffixes from higher up
the expression tree where not being included, resulting in incorrect
penalties. In some cases, this caused the formatter to believe that all
options were going to exceed 80 col, and so it failed badly by putting
[way] too much on a single line.

https://chromium-review.googlesource.com/c/chromium/src/+/2124534 is a
test run on all .gn/.gni files in Chromium. Only the first two files in
the CL are related to this change, and seem like improvements or
approximately as good.

Bug: gn:156
Change-Id: I017ef6bc2918c1c8fdbd971c1041f5069b4c26c7
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7880
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Scott Graham <scottmg@chromium.org>