Sort test file references used for indexing XCUITest test cases
By sorting the list of files we ensure that the generated file
is always the same (previously the order of the file depended
on the order where the BUILD.gn files were parsed which is not
deterministic).

Ensuring the generated project is deterministic makes it easier
to compare projects between different runs (which can be useful
when trying to change the content of the generated project to
bring it in sync which what new version of Xcode expects).

Bug: chromium/1071055
Change-Id: I62bbfd5c4112c031cb314694a69e259b5119eb0a
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8080
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>