build/gen.py: Add --link-lib=LINK_LIB option.
This adds one new command-line option to the build/gen.py
script to allow linking the final executable(s) with one or
more extra libraries.

This is mostly useful to link a custom malloc implementation
or a cpu profiling library, like libtcmalloc.a or libprofiler.a
provided by the the gperftools project [1].

For the record, using:

  build/gen.py --use-icf --use-lto --link-lib=/path/to/libtcmalloc.a

Results in a noticeably faster GN executable for both Chromium
(21% faster) and the Fuchsia build (43% faster!)

It is also possible to link against a more recent release of TCMalloc
from [2], but measurements show no noticeable performance difference
with the older release for GN workloads.

  Chromium:
    BEFORE:  Done. Made 13124 targets from 2289 files in 3593ms
    AFTER:   Done. Made 13124 targets from 2289 files in 4358ms

  Fuchsia:
    BEFORE: [...] Done. Made 47263 targets from 3819 files in 14263ms
    AFTER:  [...] Done. Made 47263 targets from 3819 files in 9923ms

[1] https://github.com/gperftools/gperftools
[2] https://github.com/google/tcmalloc

Change-Id: If21911f13a886eeae09797099a078ed1d4f789dd
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8360
Reviewed-by: Scott Graham <scottmg@chromium.org>
Commit-Queue: David Turner <digit@google.com>