Divide sources between targets and add include paths to them
Autocomplete in Xcode not works if we have all sources in one target,
at least in big projects.
So we add a few types of targets: static and dynamic libraries, with
 them, generated project better represent the internal structure of
the project. Also, we do not add all sources to one target, we spread
them between targets, where they should be.
Also, we add include paths to targets.

Change-Id: I3c81aaff5f62eff2d7a6ba1c1b14c62da6e26c27
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8380
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>