[rust-project] switch cfgs to new list schema
The `key_vlaue_cfgs: {}` schema used before didn't allow for a key
being present multiple times.  This aligns with new support in
rust-analyzer to take in the set of cfgs in the same format as
rustc produces them, as a series of key="value" pairs.

Change-Id: I1d486a670f1e2669805a55084b725798414fe5ed
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8441
Commit-Queue: Petr Hosek <phosek@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Benjamin Brittain <bwb@google.com>