Add weak_frameworks linking option
The -weak_framework linker flag adds a dynamic library dependency to the
output, but it instructs the loader to skip loading the library if it is
not present and leave the external symbols unresolved. This is typically
used when a framework is present in a new version of an SDK that is not
present on an older OS. See
https://developer.apple.com/library/archive/documentation/MacOSX/Conceptual/BPFrameworks/Concepts/WeakLinking.html
for details.

Currently users of gn implement this using a config, but due to the way
configs propagate, the -weak_link ldflags get lost at static_library
boundaries. Having first-class support for weakly linked frameworks
simplifies that.

Bug: chromium:1082896
Change-Id: Ia93bf1b67e5b9205073b8f2bf3ed9b96e3af79e4
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8464
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>