Specify deps format for actions
Setting deps format is necessary for Ninja to store the depfile in its
internal dependency database. This can significantly speed up loading
for large dependency files, which can be the case for custom language
support, e.g. Go or Dart in Fuchsia, where we emit large depfiles from
actions. This also enables the use of `ninja -t deps` tool for custom
actions.

We hardcode the format to gcc since the documentation for depfile
already explicitly specifies that "format is that of a Makefile".

Change-Id: I67dfdfe4ede8fe8d4a2a1fa3ee2cecb7c185aafd
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8722
Commit-Queue: Petr Hosek <phosek@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>