[rust-project.json]  Refactor and print each crate source root
The source root for each crate is now required by rust analyzer.
This change refactors the RustProjectJsonWriter to work in two
phases:
1) generate the full list of crates (with dependencies and cfgs)
2) write out that list of crates to the output stream

This also normalizes the output of the sysroot crates to use the
same json writing code as the rest of the crates (but sysroot
crates are still added to the crate list in a special manner).

Change-Id: I4b78ecf5444edf538b2c20088ca83372054e2098
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8720
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>