[xcode] Wrap ninja invocation for Xcode New Build System
The New Build System shipped with Xcode 10+ expects the paths
in error messages to be absolute even when invoking a custom
build script target.

Wrap the invocation of ninja in a python script that uses
regular expressions to convert the relative paths to absolute
paths.

Remove deprecated parameter --ninja-extra-args as it is hard
to support with python script (it would require parsing the
string and interpret it as a shell would do).

Bug: chromium:1094890
Change-Id: I893128f0c6744f98c8c89f1a807c83400be26545
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/9080
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>