Allow WriteCompilerBuildLine to take multiple source files
Some compilers (notably swiftc) need to be passed the full
list of source files in a single invocation, thus needs to
be passed a list of files.

Since the generated ninja output is similar in both case,
change WriteCompilerBuildLine to take a std::vector<> of
SourceFile instead of a single instance.

Bug: 121
Change-Id: I4df2976aa82b57ba5a23ad05be52bce1321100ce
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/9460
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>