Add a command line to define/overwrite "root" in ".gn"
The new command line option "--root-target" will tell GN where to find
the initial BUILD.gn. This is equivalent to define the "root" variable
in ".gn" as before. And after this CL if "root" is also defined in
".gn", this option will overwite the value defined in ".gn".

This is to achieve, to some extend, the "partial configuration"
ability. For example, if we only want to configure and build the
targets in a subfolder of Chromium like "//third_party/icu". We only
need to run

```
gn gen out/icu --root-target="//third_party/icu"
```

The motivation of this feature is to make sharing packages between
Chrome and CrOS much easier.

BUG=chromium:1133101
TEST=It can indeed only configure subfolder of chromium like
TEST="third_party/icu" and "third_party/grpc".
TEST=It can overwrite the "root" variable in ".gn" correctly.

Change-Id: If0742b409a529f34f435c6498de57667fabe7fc1
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/10140
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>