Apply path search logic more generally
The --script-executable option becomes much more portable if "python"
(or "python3") can be specified across platforms. On Windows this
requires adding a .bat or .exe extension and doing a search of the path.
The logic to do this already existed and just needed to be generalized.

Running clang-format on setup.cc adjusted a few unrelated lines.

Change-Id: If0b44449fa4f4672b5408c3c26dc21b2388dd184
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/10600
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>