Find python/python3 from dotfile entries
This is the third (and final?) step in getting gn to properly find
Python3 on Windows. The first attempt got it to translate python3.bat.
The second attempt added the ability to handle python3. This change adds
support for the directive being in the dotfile.

This is all in aid of moving Chromium (and any other projects using
gn) to Python 3.

This has been tested with the recently reverted dotfile change in
Chromium (crrev.com/c/2591588) and it successfully did an otherwise-
failing "gn gen" and a chrome build is succeeding. It was also tested
with "gn gen out\arg --script-executable=python3" and this worked.

Change-Id: I7449a989d488dd107929e997a5e7d21daadd9757
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/10900
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>