Rewriting python search logic
The various rewrites to properly search for Python on Windows whether
specified on the command-line, dotfile, or implicitly were done a bit
hurriedly and accumulated some inconsistencies and unhandled cases. This
cleans up the code to make it more obviously correct.

The initial goal was to have set_python_path call ProcessFileExtensions
but in the end that seemed to cause more problems that it solves.

Change-Id: I46133b6f015bfa5bfa77a3aa405ff906497c4045
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/10920
Reviewed-by: Brett Wilson <brettw@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>