Fix bug in analyze related to sub-configs.
Analyze had a bug where if you modified a build file containing
a config that was only used as a sub-config, analyze wouldn't
thing changes to the file would matter.

This apparently is due to sub-configs not being treated as fully
defined items in the builder (which is sensible), which led to
the analyze code not seeing them directly.

A fix for this is to walk any configs that *are* defined as part
of analyze and explicitly check them for subconfigs that might
be relevant.

Bug: 131
Change-Id: I0323c2340ccda58b341699ed221d2fd54d9d261c
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/11000
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>