[xcode] Fix detection of XCTests by Xcode 12.0+
For Xcode to detect the XCTests automatically in a project, the
source files used by the tests needs to be listed in the project.
However, if the files are listed, then Xcode tries to build them.

With version of Xcode prior to 12.0, the project was configured
so that Xcode would do nothing while trying to compile them (by
passing `--help` flag to the compiler). This stopped working with
Xcode 12.0 which started to expect the compilation to produce
some output.

Instead of configuring the compilation to be a no-op, instead
configure Xcode so that those files are not compiled by setting
a pattern matching all files for "Excluded Source File Names".
This works with all version of Xcode and both "legacy" and "new"
build system.

Bug: chromium:1103230
Change-Id: I44eb0df32ca9c8d57b2dc5b378878c5eb308bf38
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/11162
Reviewed-by: Justin Cohen <justincohen@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>