[ios] Fix example ios toolchain
If the minimum deployment target is not equal to the SDK version
(say deployment target of 14.0 using iOS 15.0 SDK) linking swift
code requires using compatibility libraries. Those are part of
the toolchain and not the SDK.

Fix the example to add the necessary path to `lib_dirs`.

Bug: none
Change-Id: Ie3ce15f851896f6623c09c88645e409b236754e8
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/11700
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>