Replace OrderedSet<> instances with UniqueVector<>.
The OrderedSet<> template has a misleading name, because its
semantics are not those of an ordered set of items, but match
completely those of UniqueVector<>. The latter is also slightly
faster and uses less memory.

This CL replaces all instances of OrderedSet<> with UniqueVector<>
ones, and removes the template definition from the source tree.

Profiling shows no noticeable difference in performance or memory usage
with the Chrome and Fuchsia "gn gen" operations.

Change-Id: Ie9b8ecfc0724d07129f69c68db5e1205b49f3970
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/11825
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>