Introduce TargetSet type
Handling sets of target pointers is a very performance-critical
operation in GN, and this is currently being done by instances
of the std::set<const Target*> type, which is slow and memory
hungry.

This CL introduces the TargetSet type as an alias to
std::set<const Target*>, and updates the source code to use it.

A future CL will change the TargetSet implementation to a better
one after this one.

Change-Id: I6e01a14d4de2d238818c8c9dd963cda5cf9e66b8
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/12581
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>