Make error information optional.
GN makes a lot of Err objects. In a normal run, all of these will be
empty (this indicates success). But an empty Err object has a Location
(1 pointer and 2 ints), a Label (4 atoms and a size_t), 2 std::vectors,
and 2 std::strings. All of these need default initializing every time
and take space.

This patch converts the "error set" state to be in a separate
dynamically created object. An empty Err is now thee size of a pointer
and is initialized as a single pointer set.

This improves GN runtime on a slow MacBook by about 4% (16235->15620ms).

Change-Id: I847944bc2be654cc391b4f161b2583959b18e4a3
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/12720
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Brett Wilson <brettw@chromium.org>