Add commands::CommandSwitches class
This class is used to model the command-line switches
that appear in various GN commands, as a set of boolean,
enum or string values, which can be swapped as a whole
globally (which might be useful when implementing a REPL
in GN in the future, since each command evaluation may
need a different set of switches then).

A future CL will modify the command implementation functions
to use the methods on this class, instead of peeking directly
into the current process' base::CommandLine value.

Bug: None
Change-Id: I240778abf9a5e4fba4760f5ca81eb84cebf2c1c8
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/12862
Reviewed-by: Brett Wilson <brettw@chromium.org>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>