[infra] Fix rpmalloc linking.
The rpmalloc configure.py script forces the '-Weverything` flag
which conflicts with more recent default C standard used by recent
Clang rolls (see associated bug with details). This results in
compilation failures.

We may be able to convince upstream to do the change themselves
but this is a temporary solution until this happens.

Bug: 278
Change-Id: I4ee10d1e10dc39b2a903cee0e1407ecc14b58ec7
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/13140
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>