Ensure all --extern crates are listed as dependencies in Ninja.
Any Rust crate B that is visible for use as a dependency of another
Rust crate A gets added to the compile of A with an `--extern` flag,
such as `--extern B`. This ensures that B is accessible from A, so that
it can use the functions and types in B.

When A makes use of something in B, it means A needs to be rebuilt when
B changes, however. This is not expressed to ninja through the
`externs` set. Currently, if A deps on B deps on C, and B is an
order-only dependency (such as a group), ninja does not know that A
should be rebuilt when C is changed. For direct dependencies,
ninja hears about them since rlibs are also collected as linkable_deps
in NinjaBinaryTargetWriter::ClassifyDependency(). But transitive
dependencies need to be propagated to the set of implicit dependencies
of a target in order to cause it to rebuild, if the dependency can be
accessed from the target.

We ensure transitive deps get a dependency relationship in ninja if
and only if they appear in `externs`, since transitive dependencies
that are not visible to a crate can not require it to be recompiled.

Bug: 282
Change-Id: I9f30274c65cccabcaed0181d7fe3b95539fb674b
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/13380
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>