Fix support for empty `script_executable` value in .gn file.
A recent CL [1] broke the Fuchsia build because the latter sets
the `script_executable` value to the empty script in its `.gn`
file. This use case is documented to let GN invoke action scripts
directly.

This fixes the issue and adds a unit-test for this use case.

[1] https://gn-review.googlesource.com/c/gn/+/13260

Bug: None
Change-Id: I2660c6de0e8f41790c60bbb072149f1d7360e7f0
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/13460
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: David Turner <digit@google.com>