[infra] Fix GN build with rpmalloc and latest Clang.
Latest Clang will now complain when the `-funit-at-a-time` optimization flag,
used by the rpmalloc build, is being used. The error message looks like:

```
.....
clang-15: error: optimization flag '-funit-at-a-time' is not supported [-Werror,-Wignored-optimization-argument]
```

This CL locally patches the rpmalloc `build/ninja/clang.py` file to add
the `-Wno-ignored-optimization-argument` flag to avoid this. This is a
temporary fix until the issue is fixed in upstream rpmalloc.

Bug: None
Change-Id: I7e3af4222a1813219843d532d61d8c2babe9b86f
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/13581
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>